Create cluster - validate availability zones count interactively
Add a validator to the selecting availability zones prompt,
to improve user experience and enable the user to fix their choices.

Related: SDA-6415
